emergencygo/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-rest-passwordreset==1.3.0
python-dotenv==1.0.0
Pillow==9.5.0 